+++ src/components/scrollContainer/index.js
@@ -1,0 +1,3 @@
+import ScrollContainer from './ScrollContainer.js';
+
+export default ScrollContainer;